refactor: unify Binance finalizers
