+++ src/finalizer/index.ts
@@ -33,8 +33,7 @@ import {
 import { ChainFinalizer, CrossChainMessage, isAugmentedTransaction } from "./types";
 import {
   arbStackFinalizer,
-  binanceL1ToL2Finalizer,
-  binanceL2ToL1Finalizer,
+  binanceFinalizer,
   cctpL1toL2Finalizer,
   cctpL2toL1Finalizer,
   heliosL1toL2Finalizer,
@@ -122,8 +121,8 @@ const chainFinalizers: { [chainId: number]: { finalizeOnL2: ChainFinalizer[]; fi
     finalizeOnL2: [],
   },
   [CHAIN_IDs.BSC]: {
-    finalizeOnL1: [binanceL2ToL1Finalizer],
-    finalizeOnL2: [binanceL1ToL2Finalizer, heliosL1toL2Finalizer],
+    finalizeOnL1: [binanceFinalizer],
+    finalizeOnL2: [heliosL1toL2Finalizer],
   },
   [CHAIN_IDs.SONEIUM]: {
     finalizeOnL1: [opStackFinalizer],
