--- src/finalizer/utils/binance/index.ts
@@ -1,2 +1,0 @@
-export * from "./l1ToL2";
-export * from "./l2ToL1";
